requirements.txt
# Core dependencies
numpy>=2.2.2
pandas>=2.2.3
geopandas>=1.0.1
pycountry==24.6.1
pydantic>=2.10.6
rasterio==1.3.10
requests>=2.32.4
scipy>=1.15.1
Shapely>=2.0.7
networkx>=3.2.1
tqdm>=4.65.0
OWSLib>=0.31.0
pydantic-settings>=2.7.1
hdx-python-api>=6.3.8
bs4==0.0.2
s3fs>=2024.12.0
psutil>=6.0.0
httpx>=0.27

# Grid systems
mercantile==1.2.1
h3>=4.2.0
s2sphere==0.2.5

# Optional: GEE Profiler (pip install "giga-spatial[gee]")
earthengine-api>=1.7.0
geemap>=0.36.6

# Optional: BigQuery / MLab (pip install "giga-spatial[bq]")
google-cloud-bigquery>=3.0.0
google-cloud-bigquery-storage>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
db-dtypes>=1.0.0

# Optional: Azure (pip install "giga-spatial[azure]")
azure-storage-blob>=12.22.0

# Optional: Snowflake (pip install "giga-spatial[snowflake]")
snowflake-connector-python>=3.0.0

# Optional: Delta Sharing (pip install "giga-spatial[delta]")
delta_sharing==1.2.0

# Optional: Database / Trino (pip install "giga-spatial[db]")
SQLAlchemy>=2.0.0
sqlalchemy-trino==0.5.0
dask>=2024.12.1

# Optional: DuckDB
duckdb==1.2.0